rename utils.py to model_utils.py and update imports
