+++ fantasytalking/nodes.py
@@ -1,7 +1,7 @@
 import os
 import torch
 import gc
-from ..utils import log
+from ..model_utils import log
 
 from accelerate import init_empty_weights
 from accelerate.utils import set_module_tensor_to_device
