+++ multitalk/nodes.py
@@ -4,7 +4,7 @@ from comfy.utils import load_torch_file, common_upscale
 from accelerate import init_empty_weights
 from accelerate.utils import set_module_tensor_to_device
 import torch
-from ..utils import log
+from ..model_utils import log
 
 
 class MultiTalkModelLoader:
